Improve rails + docker setup
This commit prevents the issue where you can't restart the container
without getting an error saying the server is already running by
setting the pid file to be outside of the shared volume. Additionally,
windows users weren't seeing their updated files reflected in the running
rails application and this appears to be caused by the evented file system
watcher. Commenting out a line in the development environment allowed
file system changes to be represented in the running application.